Split dropdown into separate file
